add api module
closes #36
